requirements.txt
streamlit
langchain
langchain-openai
langchain-huggingface
faiss-cpu
numpy
python-dotenv
python-docx
pypdf2